+++ mq_radio/engine/__init__.py
@@ -1,4 +1,9 @@
 from mq_radio.engine.audio_devices import list_audio_devices
+from mq_radio.engine.audio_router import (
+    AudioRouter,
+    apply_audio_route_from_settings,
+    get_audio_router,
+)
 from mq_radio.engine.base import EngineState, PlayoutEngine
 from mq_radio.engine.liquidsoap import LiquidsoapEngine
 from mq_radio.engine.mock_engine import MockEngine
@@ -9,4 +14,7 @@ __all__ = [
     "MockEngine",
     "LiquidsoapEngine",
     "list_audio_devices",
+    "AudioRouter",
+    "get_audio_router",
+    "apply_audio_route_from_settings",
 ]